backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-google-genai
faiss-cpu
pypdf
python-multipart
python-dotenv
numpy
pydantic